--- src/components/CreateTodo/components/index.ts
@@ -1,1 +1,0 @@
-export * from '../../ToggleableModal/ToggleableModal'
